+++ src/CRC32.h
@@ -37,15 +37,15 @@ public:
   bool getReverseOut() const { return _reverseOut; }
 
   [[deprecated("Use calc() instead")]]
-  uint8_t getCRC() const;
+  uint32_t getCRC() const;
   [[deprecated("Use setInitial() instead")]]
-  void setStartXOR(uint16_t initial) { _initial = initial; }
+  void setStartXOR(uint32_t initial) { _initial = initial; }
   [[deprecated("Use setXorOut() instead")]]
-  void setEndXOR(uint16_t xorOut) { _xorOut = xorOut; }
+  void setEndXOR(uint32_t xorOut) { _xorOut = xorOut; }
   [[deprecated("Use getInitial() instead")]]
-  uint16_t getStartXOR() const { return _initial; }
+  uint32_t getStartXOR() const { return _initial; }
   [[deprecated("Use getXorOut() instead")]]
-  uint16_t getEndXOR() const { return _xorOut; }
+  uint32_t getEndXOR() const { return _xorOut; }
   [[deprecated("Use add() with yieldPeriod instead")]]
   void enableYield() const {}
   [[deprecated("Use add() without yieldPeriod instead")]]
