+++ src/CRC64.h
@@ -37,15 +37,15 @@ public:
   bool getReverseOut() const { return _reverseOut; }
 
   [[deprecated("Use calc() instead")]]
-  uint8_t getCRC() const;
+  uint64_t getCRC() const;
   [[deprecated("Use setInitial() instead")]]
-  void setStartXOR(uint16_t initial) { _initial = initial; }
+  void setStartXOR(uint64_t initial) { _initial = initial; }
   [[deprecated("Use setXorOut() instead")]]
-  void setEndXOR(uint16_t xorOut) { _xorOut = xorOut; }
+  void setEndXOR(uint64_t xorOut) { _xorOut = xorOut; }
   [[deprecated("Use getInitial() instead")]]
-  uint16_t getStartXOR() const { return _initial; }
+  uint64_t getStartXOR() const { return _initial; }
   [[deprecated("Use getXorOut() instead")]]
-  uint16_t getEndXOR() const { return _xorOut; }
+  uint64_t getEndXOR() const { return _xorOut; }
   [[deprecated("Use add() with yieldPeriod instead")]]
   void enableYield() const {}
   [[deprecated("Use add() without yieldPeriod instead")]]
